refactor: _components 폴더로 파일 구조 리팩터링
